chore(cli): split crawler into modules; move constants; add graph export/viewer; rename to xml-indexer; rename expanders/queries; clean old files
